+++ src/openhuman/agent/tinyagents/middleware.rs
@@ -29,3 +29,4 @@ include!("middleware_part_03.rs");
 include!("middleware_part_04.rs");
 include!("middleware_part_05.rs");
 include!("middleware_part_06.rs");
+include!("middleware_part_07.rs");
